app/requirements.txt
torch==2.2.2
pillow==10.3.0
torchvision==0.17.2
requests==2.31.0
pytest==8.3.1
